changed the datatypes for the table and created a index route to get all the locaitons.

+++ server/models/hike.js
@@ -1,8 +1,8 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   const Hike = sequelize.define('Hike', {
-    latitude: DataTypes.INTEGER,
-    longitude: DataTypes.INTEGER
+    latitude: DataTypes.DECIMAL,
+    longitude: DataTypes.DECIMAL
   }, {});
   Hike.associate = function(models) {
     // associations can be defined here
